evobench-run insert: verify validity of inserted commit ids
Using a second WorkingDirectoryPool, with just 1 entry; this gets it
its own lock (running insertion and job execution processes on the
same lock would conflict), and using the pool is still useful for the
cloning logic and for the handling of errors (moving away the ~failing
working dir to make space for a fresh one, and allowing inspection of
the issue). Will use this pool both for checking commit ids, and for
polling upstream for changes.

- global_app_state_dir.rs: add `working_directory_for_polling_pool_base`

- add polling_pool.rs: WorkingDirectoryPool wrapper with currently
  just `commit_is_valid` method.

- working_directory_pool.rs:

    - process_working_directory(): make run_parameters argument
      optional, same in struct `ProcessingError`

    - add WorkingDirectoryPool.get_first() with the cloning code, as
      there was no 'generic' way to get to a working directory before