run/run_job.rs: clean up: move actions outside the working directory context
.. if they don't rely on the working directory.

Instead of contorted error capture and return, just run the actions
outside the context. This *does* need cloning the timestamp to the
outside, though. But even if
`working_directory_pool.process_working_directory()` were to reserve
the working dir for that time (lock or similar), it's OK to clone the
timestamp. The API is fine to allow it to be cloned, nothing can go
wrong with that.