evobench-run daemon: fix "run grep-diff functionality via (optional) configuration"
I had accidentally made the `log_extracts` configuration a part of the
`Key` (if that were up to date), or at least the key hashing of the
already-inserted table.  It doesn't belong there since it is just
configuration for *derived* values, not a primary benchmarking input.

run/config.rs:

- move `log_extracts` field from `BenchmarkingCommand` to
  `BenchmarkingTarget`
- store `targets` in `RunConfig`

run/run_job.rs:

- add docs
- add JobRunner.run_config field
- BenchmarkingCommand.target_name is now what we return from the
  working directory context
- look up target_name in the RunConfig

bin/evobench-run.rs:

- pass run_config to
- no log_extracts in