evobench run: fix: do not use the polling pool for startup check
The lock on the pool led to errors when `evobench-run poll` is run at
the same time. That error which is on purpose because no more than one
instance of the `poll` and `run` actions should ever be executed at
the same time. Thus, change to use the proper pool for `run` actions.