evobench-run run: take a new global lock, separate from working directory lock
This because the working directory lock needs to become fine grained
so that multiple programs can work on it concurrently.

Now it's clear what each lock is for, too.

This also adds a configuration change, and a default configuration
entry. The configuration change is backwards compatible thanks to the
new field being optional, though.