evobench-run wd: add `recycle` subcommand
It is OK to use `do_mark`, it won't fail: the encoded logic is to
complain when taking a dir *away* from the daemon; but this is giving
it to it, which is fine.  We just need to signal it.

Using LazyResult to avoid loading the signalling file in the other
cases (but share it between the subcommands that do).